+++ src/main/java/com/example/skolaback/model/entity/ContestApplication.java
@@ -1,7 +1,8 @@
 package com.example.skolaback.model.entity;
 
-import jakarta.persistence.*;
 import lombok.Data;
+
+import javax.persistence.*;
 
 @Data
 @Entity
@@ -10,7 +11,7 @@ public class ContestApplication {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, nullable = false)
-    private Integer id;
+    private long id;
 
     @ManyToOne
     @JoinColumn(name = "contest_id")
